This is the magic jar.
This was originally compiled using an early version of XCode. The XCode project file format has changed and so has the java comiling stuff. As non-java developers, the build error is, um, opaque:
"Jam is deprecated and has been removed; targets that use Jam must be upgraded to native targets. For more information on doing this, consult the Xcode documentation."

Documentation and verious folk online all say that we need to move to ant based build instead, and then I end up in a bunch of twisty java build documentation, all alike...

So, I'm just using this xml2rfc.jar file from an earlier version an praying that it doens't ever need to be rebuilt...