Updates the build system to make two distinct targets, a build for 5.x.x and a build for 6.x.x
This is needed because otherwise the xml2rfc.xxe_addon that gets baked into the zip file refers to either
version 5.x.x or version 6.x.x. This is similar to the issue that created the lst set of "support multipl versions" commits,
but *only* affects people testing by copying the zip file locally. As this is usually me (for testing), and I have a really bad
memory, this is important....

This now builds a ./dist/5/ and a ./dist/6/ tree. It also does variable substitution for the major version of XMLMind and version of 
the plugin.

"This time for sure, rocky"